chore: Splitting files to support groups on members page in EE (#18085)
* splitted files to support groups on members page in EE

* updated an import

* minor change
