+++ app/client/src/ce/pages/workspace/WorkspaceInviteUsersForm.tsx
@@ -61,7 +61,7 @@ import ManageUsers from "pages/workspace/ManageUsers";
 import { ScrollIndicator } from "design-system";
 import UserApi from "@appsmith/api/UserApi";
 import { Colors } from "constants/Colors";
-import { fetchWorkspace } from "actions/workspaceActions";
+import { fetchWorkspace } from "@appsmith/actions/workspaceActions";
 import { useHistory } from "react-router-dom";
 import { Tooltip } from "@blueprintjs/core";
 import { isEllipsisActive } from "utils/helpers";
