+++ app/client/src/ce/reducers/uiReducers/workspaceReducer.ts
@@ -8,7 +8,7 @@ import {
   WorkspaceRole,
   Workspace,
   WorkspaceUser,
-} from "constants/workspaceConstants";
+} from "@appsmith/constants/workspaceConstants";
 
 export const initialState: WorkspaceReduxState = {
   loadingStates: {
