+++ app/client/src/components/editorComponents/ActionCreator/FieldConfig.ts
@@ -6,7 +6,7 @@ import {
   RESET_CHILDREN_OPTIONS,
   ViewTypes,
 } from "./constants";
-import { ALERT_STYLE_OPTIONS } from "../../../ce/constants/messages";
+import { ALERT_STYLE_OPTIONS } from "@appsmith/constants/messages";
 import { ActionType, AppsmithFunctionConfigType, FieldProps } from "./types";
 import {
   enumTypeGetter,
