+++ app/client/src/ee/actions/workspaceActions.ts
@@ -1,0 +1,1 @@
+export * from "ce/actions/workspaceActions";
