+++ app/client/src/ee/constants/workspaceConstants.ts
@@ -1,0 +1,1 @@
+export * from "ce/constants/workspaceConstants";
