+++ app/client/src/ee/pages/workspace/Members.tsx
@@ -1,0 +1,3 @@
+export * from "ce/pages/workspace/Members";
+import { default as CE_Members } from "ce/pages/workspace/Members";
+export default CE_Members;
