+++ app/client/src/sagas/ApplicationSagas.tsx
@@ -54,7 +54,7 @@ import {
 } from "@appsmith/constants/messages";
 import { Toaster, Variant } from "design-system";
 import { APP_MODE } from "entities/App";
-import { Workspace, Workspaces } from "constants/workspaceConstants";
+import { Workspace, Workspaces } from "@appsmith/constants/workspaceConstants";
 import { AppIconName } from "design-system";
 import { AppColorCode } from "constants/DefaultTheme";
 import {
