+++ app/client/src/selectors/applicationSelectors.tsx
@@ -9,7 +9,7 @@ import {
   WorkspaceDetails,
 } from "@appsmith/constants/ReduxActionConstants";
 import Fuse from "fuse.js";
-import { Workspaces } from "constants/workspaceConstants";
+import { Workspaces } from "@appsmith/constants/workspaceConstants";
 import { GitApplicationMetadata } from "api/ApplicationApi";
 import {
   isPermitted,
