+++ app/client/src/utils/helpers.tsx
@@ -10,7 +10,7 @@ import {
   JAVASCRIPT_KEYWORDS,
 } from "constants/WidgetValidation";
 import { get, set, isNil, has } from "lodash";
-import { Workspace } from "constants/workspaceConstants";
+import { Workspace } from "@appsmith/constants/workspaceConstants";
 import {
   isPermitted,
   PERMISSION_TYPE,
